requirements.txt
gunicorn
alabaster==0.7.12
asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.9.1
bootstrap4==0.1.0
cairocffi==1.2.0
CairoSVG==2.5.0
boto3==1.15.6
botocore==1.18.6
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cssselect2==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-allauth==0.42.0
django-bootstrap4==2.2.0
django-braces==1.14.0
django-ckeditor==5.9.0
django-classy-tags==1.0.0
django-cms==3.7.3
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-discover-runner==1.0
django-extensions==3.0.5
django-filer==1.7.1
django-filter==2.3.0
django-formtools==2.2
django-js-asset==1.2.2
django-mailchimp==0.1.30
django-model-utils==4.0.0
django-mptt==0.11.0
django-paypal==1.0.0
django-polymorphic==2.1.2
django-realtime==1.1
django-rest-framework==0.1.0
django-restframework==0.0.1
django-sekizai==1.1.0
django-star-ratings==0.9.1
django-storages==1.9.1
django-storages-redux==1.3.3
django-tenant-schemas==1.10.0
django-tenants==3.1.0
django-treebeard==4.3.1
django-utils-six==2.0
django-weasyprint==1.0.2
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.9.1
djangocms-video==2.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
easy-thumbnails==2.7
fpdf==1.7.2
html5lib==1.1
httpie==2.1.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.7.0
Jinja2==2.11.2
jmespath==0.10.0
ldif3==3.1.1
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
ordered-set==4.0.1
packaging==20.3
Pillow==7.2.0
# psycopg2==2.8.5
# psycopg2-binary==2.8.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
Pyphen==0.10.0
python-card-me==0.9.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snowballstemmer==2.0.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
static3==0.7.0
surlex==0.2.0
swapper==1.1.2.post1
tinycss2==1.1.0
tzlocal==2.1
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.9
WeasyPrint==52.2
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.1.0
